@@ -15,16 +15,16 @@
   ],
   "dependencies": {
     "@angular-slider/ngx-slider": "^2.0.4",
-    "@angular/animations": "10.2.5",
+    "@angular/animations": "11.2.14",
     "@angular/cdk": "^10.2.7",
-    "@angular/common": "10.2.5",
-    "@angular/compiler": "10.2.5",
-    "@angular/core": "10.2.5",
-    "@angular/forms": "10.2.5",
+    "@angular/common": "11.2.14",
+    "@angular/compiler": "11.2.14",
+    "@angular/core": "11.2.14",
+    "@angular/forms": "11.2.14",
     "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "10.2.5",
-    "@angular/platform-browser-dynamic": "10.2.5",
-    "@angular/router": "10.2.5",
+    "@angular/platform-browser": "11.2.14",
+    "@angular/platform-browser-dynamic": "11.2.14",
+    "@angular/router": "11.2.14",
     "@babel/preset-env": "^7.22.5",
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
@@ -63,10 +63,10 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1002.3",
-    "@angular/cli": "10.2.3",
-    "@angular/compiler-cli": "10.2.5",
-    "@angular/localize": "^10.2.5",
+    "@angular-devkit/build-angular": "~0.1102.19",
+    "@angular/cli": "11.2.19",
+    "@angular/compiler-cli": "11.2.14",
+    "@angular/localize": "^11.2.14",
     "@babel/traverse": "^7.22.10",
     "@ngtools/webpack": "10.2.3",
     "@openapitools/openapi-generator-cli": "0.0.14-4.0.2",
